requirements.txt
black==24.8.0
click==8.1.7
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.1
pygame-ce==2.5.1
